+++ exchangelib/__init__.py
@@ -9,7 +9,7 @@ from .credentials import DELEGATE, IMPERSONATION, Credentials, ServiceAccount
 from .ewsdatetime import EWSDate, EWSDateTime, EWSTimeZone, UTC, UTC_NOW
 from .extended_properties import ExtendedProperty, ExternId
 from .folders import SHALLOW, DEEP
-from .items import CalendarItem, Contact, Message, Task
+from .items import CalendarItem, Contact, DistributionList, Message, PostItem, Task
 from .properties import Body, HTMLBody, ItemId, Mailbox, Attendee, Room, RoomList
 from .restriction import Q
 from .transport import BASIC, DIGEST, NTLM
@@ -23,7 +23,7 @@ __all__ = [
     'DELEGATE', 'IMPERSONATION', 'Credentials', 'ServiceAccount',
     'EWSDate', 'EWSDateTime', 'EWSTimeZone', 'UTC', 'UTC_NOW',
     'ExtendedProperty',
-    'CalendarItem', 'Contact', 'Message', 'Task',
+    'CalendarItem', 'Contact', 'DistributionList', 'Message', 'PostItem', 'Task',
     'ItemId', 'Mailbox', 'Attendee', 'Room', 'RoomList', 'Body', 'HTMLBody',
     'Q',
     'SHALLOW', 'DEEP',